+++ src/main/java/roomescape/config/PaymentConfig.java
@@ -1,16 +1,21 @@
 package roomescape.config;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
+import roomescape.payment.client.TossPayRestClient;
 
 @Configuration
 public class PaymentConfig {
 
+    @Value("${payment.secret-key}")
+    private String paymentSecretKey;
+
     @Bean
-    public RestClient restClient() {
-        return RestClient.builder()
+    public TossPayRestClient tossPayRestClient() {
+        return new TossPayRestClient(paymentSecretKey, RestClient.builder()
                 .baseUrl("https://api.tosspayments.com")
-                .build();
+                .build());
     }
 }
