+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -50,7 +50,7 @@ public class ReservationService {
             ReservationSaveRequest reservationSaveRequest,
             LoginMember loginMember
     ) {
-        paymentService.payment(PaymentRequest.from(reservationSaveRequest));
+        paymentService.pay(PaymentRequest.from(reservationSaveRequest));
         Reservation reservation = createValidatedReservationOfStatus(reservationSaveRequest, loginMember, ReservationStatus.SUCCESS);
         validateDuplicatedReservationSuccess(reservation);
         Reservation savedReservation = reservationRepository.save(reservation);
